Move processStats into stats folder
