+++ src/lib/util.ts
@@ -1,0 +1,6 @@
+import moment from "moment";
+
+export function convertFrameCountToDurationString(frameCount: number): string {
+  const duration = moment.duration(frameCount / 60, "seconds");
+  return moment.utc(duration.as("milliseconds")).format("m:ss");
+}
